Moved logic from Proxy to ProjectControl.
Now the projectControl handles identifiying the type of file and
if a cif header is needed will provide this before calling the
actual loading which still occurs in the Proxy layer.

Unit tetsts for proxy updated, still need to add tests for the
part in ProjectControl.

Behavior of the app is unchanged.
